+++ src/settings/CustomFieldsSettings.js
@@ -1,4 +1,7 @@
-import { Route, Switch } from 'react-router-dom';
+import {
+  Route,
+  Switch,
+} from 'react-router-dom';
 
 import { CUSTOM_FIELDS_ORDERS_BACKEND_NAME } from '@folio/stripes-acq-components';
 import { useStripes } from '@folio/stripes/core';